Move some more color functions into PaletteState
